Slim DrivAerML trainer entrypoint
